+++ src/config/index.ts
@@ -37,5 +37,4 @@ export const Config = {
     usdcPoolPrivateKey: loadEnvironmentVariable('USDC_POOL_PRIVATE_KEY'),
   },
   clientHost: loadEnvironmentVariable('CLIENT_HOST'),
-  host: loadEnvironmentVariable('HOST'),
 };
